systemd-sleep: use swaps in priority order
In situations where hibernation is requested but resume= and
resume_offset= kernel parameters are not configured, systemd
will attempt to locate a suitable swap location by inspecting
/proc/swaps. This change will use the first suitable swap with
the highest configured priority.

+++ src/shared/sleep-config.h
@@ -29,7 +29,7 @@ int read_fiemap(int fd, struct fiemap **ret);
 #if 0 /// UNNEEDED by elogind
 int parse_sleep_config(SleepConfig **sleep_config);
 #endif // 0
-int find_hibernate_location(char **device, char **type, size_t *size, size_t *used);
+int find_hibernate_location(char **device, char **type, uint64_t *size, uint64_t *used);
 
 #if 0 /// elogind has to transport its manager instance
 int can_sleep(const char *verb);
